% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VGQ_scoring.R
\name{get_vgq_scores_latl}
\alias{get_vgq_scores_latl}
\title{Returns Video Game Player Category}
\usage{
get_vgq_scores_latl(raw.df)
}
\arguments{
\item{raw.df}{Data frame with at least one column named "Participant_ID" with the unique
IDs for each participant and the columns from the VGQ with the names
following the pattern from the Qualtrics questionnaire from the Mini-copter
study (from LATL), see here : https://uwmadison.co1.qualtrics.com/survey-builder/SV_7UGlBut4jMHOb4O/edit}
}
\value{
A data frame with the player cateogry in the column "player_category" and the item columns renamed according to the mapping above.
}
\description{
Given a data frame where each row is a participant and with the answers to the
VGQ in the columns, this functions returns the category of each participant in
terms of the VGQ (https://www.unige.ch/fapse/brainlearning/vgq/), 2022 version.
}
\details{
The output data frame has columned named following the template : <vg category>_<time period>_<measure>.

<vg catgory> can take the values "fps", "arpg", "sports", "rts", "tbrpg", "music", "tbs", or "other"
<time period> can take the values "past.year" or "before.past.year"
<measure> can take the values "hours", "small.screen", or "touch.screen"

 Requires : pracma, dplyr
}
\examples{
raw.df <- read_excel('video_game_scores.xlsx')
scored.df <- get_vgq_scores_latl(raw.df)

}
