% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latltools-package.R
\docType{package}
\name{latltools-package}
\alias{latltools}
\alias{latltools-package}
\title{latltools: Collection of Tools for LATL}
\description{
This package is collection of tools used in the LATL. For now only contains the vgq scoring function. More to come.
}
\keyword{internal}
