% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_vgq_col_names_better_latl.R
\name{make_vgq_col_names_better_latl}
\alias{make_vgq_col_names_better_latl}
\title{Uniformizes the names of the columns in a dataset from the VGQ}
\usage{
make_vgq_col_names_better_latl(raw.df)
}
\arguments{
\item{raw.df}{Data frame with at least one column named "participant_id" with the unique
IDs for each participant and the columns from the VGQ with the names
following the pattern from the Qualtrics questionnaire from the Mini-copter
study (from LATL), see here : https://uwmadison.co1.qualtrics.com/survey-builder/SV_7UGlBut4jMHOb4O/edit}
}
\value{
A data frame with the columns renamed
}
\description{
Converts the column names from the LATL VGQ Qualtrics questionnaire to a
uniformized model that follows this template : <vg category>_<time period>_<measure>.
}
\details{
<vg catgory> can take the values "fps", "arpg", "sports", "rts", "tbrpg", "music", "tbs", or "other"
<time period> can take the values "past.year" or "before.past.year"
<measure> can take the values "hours", "small.screen", or "touch.screen"
}
\examples{
scored.df <- make_vgq_col_names_better_latl(raw.df)

}
\keyword{internal}
