% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_plot.R
\name{save_plot}
\alias{save_plot}
\title{Saves an ggplot object as png, eps, and rds.}
\usage{
save_plot(plot, file_name_full_no_ext, width = 7, height = 7, units = "in")
}
\arguments{
\item{plot}{plot object from ggplot}

\item{file_name_full_no_ext}{name of the file including the path to the folder
where you want to save the figure, without the extension (e.g. png, eps, jpg,
rds, etc...) as these are added within the function}

\item{width}{width of picture (see ggsave)}

\item{height}{width of picture (see ggsave)}

\item{units}{units of width and height of picture (see ggsave)}
}
\value{
A data frame with the player cateogry in the column "player_category" and the item columns renamed according to the mapping above.
}
\description{
Given a plot object output from ggplot function, this function saves the plot
as a png (pixel map), eps (vectorized format), and rds (R object) files.
}
\examples{
# saves my_plot in 3 files my_awesome_plot.png, and my_awesome_plot.rds,
# my_awesome_plot.eps in the folder path/to/figure/directory.
my_plot <- ggplot() + ...
save_plot(plot = my_plot,
}
